The documentation hass been perfectionated
Fixed a bug that occured when setWeight method was called. The average wasnt being automatically updated.

changed return type of calcStudentsAvg to void since it is only used to asign final scores, hence I made it do so automatically in the methos. plus, it simplifies the implementation of setWeight in Assignment